poc snaphack
